+++ tests/test_utils.py
@@ -4,12 +4,14 @@
 
 from unittest import TestCase
 
+import numpy as np
 import pytest
 import xarray as xr
 
 from tests.helpers import make_s2_msi
 from xarray_eopf.utils import (
     NameFilter,
+    NameTypeMapping,
     assert_arg_is_instance,
     assert_arg_is_one_of,
     get_data_tree_item,
@@ -101,3 +103,40 @@ class NameFilterTest(TestCase):
         self.assertEqual(
             ["ernie", "emmie"], list(f.filter(["bibo", "ernie", "bert", "emmie"]))
         )
+
+
+class NameTypeMappingTest(TestCase):
+    def test_new_with_data_is_none(self):
+        ntm = NameTypeMapping.new("spline_orders", 3, None)
+        self.assertEqual(3, ntm.get("scl", np.integer))
+        self.assertEqual(3, ntm.get("b03", np.floating))
+        self.assertEqual(3, ntm.get("b02", np.floating))
+
+    def test_new_with_data_is_default(self):
+        ntm = NameTypeMapping.new("spline_orders", 3, 2)
+        self.assertEqual(2, ntm.get("scl", np.uint8))
+        self.assertEqual(2, ntm.get("b03", np.float32))
+        self.assertEqual(2, ntm.get("b02", np.int32))
+
+    def test_new_with_data_is_dict(self):
+        ntm = NameTypeMapping.new(
+            "spline_orders", 3, {0: ["scl"], 1: [np.floating, "cld"]}
+        )
+        self.assertEqual(0, ntm.get("scl", np.uint8))
+        self.assertEqual(1, ntm.get("b03", np.float32))
+        self.assertEqual(1, ntm.get("cld", np.int16))
+        self.assertEqual(3, ntm.get("b06", np.int32))
+
+    # noinspection PyMethodMayBeStatic
+    def test_new_fail(self):
+        with pytest.raises(
+            TypeError,
+            match=r"spline_orders\[1\]: name or dtype expected, but got value 3",
+        ):
+            NameTypeMapping.new("spline_orders", 3, {1: [np.floating, "cld", 3]})
+
+        with pytest.raises(
+            TypeError, match="agg_methods: keys must be of type str, but got key 1"
+        ):
+            # noinspection PyTypeChecker
+            NameTypeMapping.new("agg_methods", "´mean", {1: [np.floating, "cld"]})