+++ xarray_eopf/spatial.py
@@ -3,45 +3,59 @@
 #  https://opensource.org/license/apache-2-0.
 
 from collections.abc import Hashable, Mapping
-from typing import Any, Literal, TypeAlias
+from math import ceil
+from typing import Callable, Literal, TypeAlias
 
+import dask.array as da
 import dask_image.ndinterp as ndinterp
 import numpy as np
 import xarray as xr
 
-from xarray_eopf.utils import timeit
+import xarray_eopf.coarsen as xec
+
+from .utils import NameTypeMapping, timeit
+
+ALL = slice(None)
 
 _DEBUG = False
 
 AggMethod: TypeAlias = Literal[
-    "all",
-    "any",
+    "center",
     "count",
+    "first",
+    "last",
     "max",
     "mean",
     "median",
+    "mode",
     "min",
     "prod",
     "std",
     "sum",
     "var",
 ]
+AggMethods: TypeAlias = AggMethod | dict[AggMethod, list[str | np.dtype]]
 
-AGG_METHODS = (
-    "all",
-    "any",
-    "count",
-    "max",
-    "mean",
-    "median",
-    "min",
-    "prod",
-    "std",
-    "sum",
-    "var",
-)
+AggFunction: TypeAlias = Callable[[np.ndarray, tuple[int, ...] | None], np.ndarray]
+
+AGG_METHODS: dict[AggMethod, AggFunction] = {
+    "center": xec.center,
+    "count": np.count_nonzero,
+    "first": xec.first,
+    "last": xec.last,
+    "prod": np.nanprod,
+    "max": np.nanmax,
+    "mean": xec.mean,
+    "median": xec.median,
+    "min": np.nanmin,
+    "mode": xec.mode,
+    "std": xec.std,
+    "sum": np.nansum,
+    "var": xec.var,
+}
 
 SplineOrder: TypeAlias = Literal[0, 1, 2, 3]
+SplineOrders: TypeAlias = SplineOrder | dict[SplineOrder, list[str | np.dtype]]
 
 SPLINE_ORDERS = 0, 1, 2, 3
 
@@ -75,67 +89,76 @@ def get_ref_var_name(variables: Mapping[Hashable, xr.DataArray]) -> Hashable | N
 def rescale_spatial_vars(
     variables: Mapping[Hashable, xr.DataArray],
     ref_var_name: Hashable | None = None,
-    agg_method: AggMethod | dict[Hashable, AggMethod] | None = None,
-    spline_order: SplineOrder | dict[Hashable, SplineOrder] | None = None,
-    eps: float = 1e-5,
+    spline_orders: SplineOrders | None = None,
+    agg_methods: AggMethods | None = None,
+    eps: float = 1e-6,
 ) -> Mapping[Hashable, xr.DataArray]:
     spatial_variables = get_spatial_vars(variables)
     ref_var_name = ref_var_name or get_ref_var_name(spatial_variables)
     ref_var = spatial_variables[ref_var_name]
     ref_spatial_shape = ref_var.shape[-2:]
+    spline_orders = NameTypeMapping.new("spline_orders", 3, spline_orders)
+    agg_methods = NameTypeMapping.new("agg_methods", "mean", agg_methods)
+
     rescaled_variables = {}
 
     def format_factor(factor):
         return f"{factor:.6f}".rstrip("0").rstrip(".")
 
     for var_name, var in spatial_variables.items():
-        var_attrs = dict(var.attrs)
         ref_y_dim, ref_x_dim = ref_var.dims[-2:]
         y_dim, x_dim = var.dims[-2:]
         spatial_shape = var.shape[-2:]
         if spatial_shape != ref_spatial_shape:
-            y_scale = spatial_shape[0] / ref_spatial_shape[0]
-            x_scale = spatial_shape[1] / ref_spatial_shape[1]
-            is_down_sampling = x_scale + eps > 1.0
-            is_integer = np.issubdtype(var.dtype, np.integer)
-            if is_down_sampling:
-                method = get_agg_method(var_name, agg_method, is_categorical=is_integer)
-                isx = int(x_scale)
-                isy = int(y_scale)
-                if abs(isx - x_scale) < eps and abs(isy - y_scale) < eps:
-                    with timeit(f"down-sampling {var_name!r}", silent=not _DEBUG):
-                        coarsened = var.coarsen(**{y_dim: isy, x_dim: isx})
-                        rescaled_data = getattr(coarsened, method)()
-                    var_attrs["history"] = (
-                        f"Down-sampling by factors"
-                        f" {ref_x_dim}={isx} and {ref_y_dim}={isy}"
-                        f" using aggregation method {method!r}."
-                    )
-                else:
-                    # TODO: implement down-sampling for non-integer factors
-                    raise NotImplementedError(
-                        "Down-sampling only implemented for integer factors"
-                    )
-            else:
+            y_size, x_size = spatial_shape
+            target_y_size, target_x_size = ref_spatial_shape
+            x_scale = x_size / target_x_size
+            y_scale = y_size / target_y_size
+            x_window_size = ceil(x_scale)
+            y_window_size = ceil(y_scale)
+            target_x_size = x_window_size * target_x_size
+            target_y_size = y_window_size * target_y_size
+            x_scale = x_size / target_x_size
+            y_scale = y_size / target_y_size
+            history = var.attrs.get("history", "")
+            rescaled_data = da.asarray(var.data)
+            if abs(x_scale - 1) > eps or abs(y_scale - 1) > eps:
                 factors = (var.ndim - 2) * (1,) + (y_scale, x_scale)
                 matrix = np.diag(factors)
-                order = get_spline_order(
-                    var_name, spline_order, is_categorical=is_integer
-                )
-                with timeit(f"up-sampling {var_name}!r", silent=not _DEBUG):
+                order = spline_orders.get(var_name, var.dtype)
+                with timeit(f"Upscaling {var_name}!r", silent=not _DEBUG):
                     rescaled_data = ndinterp.affine_transform(
-                        var.data,
+                        rescaled_data,
                         matrix,
                         order=order,
-                        output_chunks=ref_var.chunks,
-                        output_shape=var.shape[:-2] + ref_spatial_shape,
+                        output_shape=var.shape[:-2] + (target_y_size, target_x_size),
                     )
-                var_attrs["history"] = (
-                    f"Up-sampling by factors"
-                    f" {ref_x_dim}={format_factor(x_scale)} and"
-                    f" {ref_y_dim}={format_factor(y_scale)}"
-                    f" using spline interpolation of order {order}."
+                x_sf = format_factor(x_scale)
+                y_sf = format_factor(y_scale)
+                s = f" {x_sf}" if x_sf == y_sf else f"s {x_sf} and {y_sf}"
+                history += (
+                    f"Upscaling dimensions {ref_x_dim!r} and {ref_y_dim!r} by"
+                    f" scale factor{s} using spline interpolation of order {order};\n"
                 )
+            if x_window_size > 1 or y_window_size > 1:
+                method = agg_methods.get(var_name, var.dtype)
+                with timeit(f"Downscaling {var_name!r}", silent=not _DEBUG):
+                    reduction: AggFunction = AGG_METHODS[method]
+                    ndim = rescaled_data.ndim
+                    # noinspection PyTypeChecker
+                    rescaled_data = da.coarsen(
+                        reduction,
+                        rescaled_data,
+                        {ndim - 1: y_window_size, ndim - 2: x_window_size},
+                    )
+                x_ws = x_window_size
+                y_ws = y_window_size
+                s = f" {x_ws}" if x_ws == y_ws else f"s {x_ws} and {y_ws}"
+                history += (
+                    f"Downscaling dimensions {ref_x_dim!r} and {ref_y_dim!r} by"
+                    f" window size{s} using aggregation method {method!r};\n"
+                )
+
             rescaled_data = rescaled_data.astype(var.dtype)
             coords = dict(var.coords)
             coords.pop(x_dim, None)
@@ -147,7 +170,7 @@ def rescale_spatial_vars(
                 data=rescaled_data,
                 dims=var.dims[:-2] + ref_var.dims[-2:],
                 name=var.name,
-                attrs=var_attrs,
+                attrs={**var.attrs, "history": history},
             ).chunk({ref_x_dim: ref_var.chunks[-1], ref_y_dim: ref_var.chunks[-2]})
             for enc_name in ("chunks", "preferred_chunks"):
                 if enc_name in ref_var.encoding:
@@ -156,42 +179,3 @@ def rescale_spatial_vars(
     all_variables = {**variables, **rescaled_variables}
     sorted_var_names = sorted(all_variables)
     return {var_name: all_variables[var_name] for var_name in sorted_var_names}
-
-
-def get_spline_order(
-    var_name: Hashable, spline_order: Any, is_categorical: bool = False
-) -> SplineOrder:
-    order = (
-        spline_order.get(var_name) if isinstance(spline_order, dict) else spline_order
-    )
-    if order is None:
-        if is_categorical:
-            order = 0  # nearest
-        else:
-            order = 3  # cubic
-    if order not in SPLINE_ORDERS:
-        raise ValueError(f"Unknown spline order: {order}")
-    return order
-
-
-def get_agg_method(
-    var_name: Hashable, agg_method: Any = None, is_categorical: bool = False
-) -> AggMethod:
-    method = agg_method.get(var_name) if isinstance(agg_method, dict) else agg_method
-    if method is None:
-        if is_categorical:
-            # TODO: down-sample categorical data
-            #  according to Sentinel-2 products, SCL should be down-sampled
-            #  using nearest-neighbour. This would be method="center".
-            #  In general, the most natural solution is to take the
-            #  most frequent value. This would be method="mode".
-            #  The only value preserving methods currently available in
-            #  xarray are "min" and "max".
-            #
-            method = "max"
-        else:
-            # For floating point we block-average
-            method = "mean"
-    if method not in AGG_METHODS:
-        raise ValueError(f"Unknown aggregation method: {method}")
-    return method