+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## Changes in 0.1.0 (under development)
 
+* Added initial analysis mode for Sentinel-2 L1C and L2A products.
+  Using analysis mode (the default), users can open deeply nested 
+  products as a flat xarray dataset with all bands resampled to a single
+  resolution. 
 * Added CI for unit and integration tests
 * Added CodeCov report
 * Added notebook examples for accessing Sentinel-1 and Sentinel-2 using the